refactor: extract useKeyboardShortcut to shared hook
- Create shared use-keyboard-shortcut hook with callback ref pattern
- Replace duplicate implementations in mode-switcher and feeback
- Fix RefObject type in useOnClickOutside to accept nullable refs
- Stabilize callback dependencies in useEventRounds with useCallback
